requirements.txt
ultralytics
opencv-python
gradio
pandas
numpy
torch
torchvision
Pillow
PyYAML
paddlepaddle
paddleocr
easyocr
scipy
filterpy
yt-dlp
requests
PyQt6
matplotlib
